Refactor Packages and Class Names
