+++ app/src/main/java/com/survivalcoding/todolist/data/TodoRepository.kt
@@ -1,4 +1,6 @@
-package com.survivalcoding.todolist
+package com.survivalcoding.todolist.data
+
+import com.survivalcoding.todolist.model.Todo
 
 class TodoRepository {
