+++ app/src/main/java/com/survivalcoding/todolist/presentation/add/AddActivity.kt
@@ -1,14 +1,15 @@
-package com.survivalcoding.todolist
+package com.survivalcoding.todolist.presentation.add
 
 import android.app.Activity
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import androidx.appcompat.app.AppCompatActivity
+import com.survivalcoding.todolist.R
 
-class WriteActivity : AppCompatActivity() {
+class AddActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_write)
+        setContentView(R.layout.activity_add)
 
         val finishBt = findViewById<Button>(R.id.finishBt)
         finishBt.setOnClickListener {
