+++ app/src/main/java/com/survivalcoding/todolist/presentation/main/MainViewModel.kt
@@ -1,8 +1,9 @@
-package com.survivalcoding.todolist
+package com.survivalcoding.todolist.presentation.main
 
 import androidx.lifecycle.ViewModel
+import com.survivalcoding.todolist.data.TodoRepository
 
-class MainActivityViewModel(): ViewModel() {
+class MainViewModel: ViewModel() {
 
     private val todoRepository = TodoRepository()
