+++ app/src/main/java/com/survivalcoding/todolist/presentation/main/adapter/TodoDiffItemCallback.kt
@@ -1,6 +1,7 @@
-package com.survivalcoding.todolist
+package com.survivalcoding.todolist.presentation.main.adapter
 
 import androidx.recyclerview.widget.DiffUtil
+import com.survivalcoding.todolist.model.Todo
 
 object TodoDiffItemCallback: DiffUtil.ItemCallback<Todo>() {
     override fun areItemsTheSame(oldItem: Todo, newItem: Todo): Boolean {
