+++ app/src/main/res/layout/activity_add.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WriteActivity">
+    tools:context=".presentation.add.AddActivity">
 
     <Button
         android:id="@+id/finishBt"
